Add files via upload
this is the buggy code. When I click on the nodes, force directed appears. But if I click again, second one disappears without deleting the first one. I have tried removing it in the beginning of force directed function and onclick function, but then the arc diagram also gets removed.